Allow classes to set document_type.
This functionality used to exist in SearchKick but no longer does.
In some cases of STI it may be desireable to have child classes use
the same mapping set as the parent class, rather then setting new
mappings for each subclass.  This allows you to easily set the
document_type method on a class in order to allow all STI tables
to use the same mappings.