Support for aliasing old and new indices when reindexing
The current searchkick implementation makes it a bit difficult if you want to reindex, but still allow reads and writes. This change creates seperate "read" and "write" aliases. While reindexing the read alias will go only to the old index, but the write alias will go to the new and the old. So in your save you can iterate through the indexes attached to `model_write` and write to both, to ensure your results do not get stale.