Resolve lazy rescue
Rescuing from find on an ActiveRecord model is fine, but
we can skip the slow rescue flow by using find_by instead.
Since we were rescuing and returning nil, find_by will
provide the same behavior.

Notably, one possible regression here is that if find_by
fails for some other reason, such as no connection to the
database, this will end up raising that exception instead
of simply returning nil as the old solution does. Arguably
this is better, though, as we wouldn't want to remove a
record from the search index simply because we couldn't
connect to the database at the right moment.