Fix infinite recursion on multi search
When we retry a multi search, we should not perform subsequent retries,
otherwise we'll end in an inifite recursion if we don't have enough
matches.

In this commit, a new flag is added to mark the execution of a retry.
Also since we have a local variable and a method with the same name
`queries`, a rename was applied to the method argument to improve the
readability.