Update Readme.md Heroku / foundelasticsearch
For Heroku with plugin foundelasticsearch and last version.
I did the update today and got Faraday SSL Error for each request after the deploy, I don't know what changed in any dependency or even in searchkick but adding the port resolve the issue. It was working fine with earlier versions of`elasticsearch-ruby` and deps.

Maybe there is the same issue with bonsai addon, i don't know. Maybe someone should verify.