QAOPS-5999 Updates searchkick:reindex:all autoload
We are currently receiving a subset of searchkick models to reindex when
invoking the searchkick:reindex:all rake task. This is due to
autoloading changes. In order to resolve the issue, we're updating the
rake task to handle for zeitwerk to ensure all models will be loaded and
subsequently reindexed.