Simplify how we create the erroring query
When trying to exceed the clause limit, one complication was that there
is a different default clause limit for Elasticsearch vs OpenSearch, and
another issue is that these constructed queries with a large number of
clauses were very slow, which in a best case was enough to slow down the
test suite and in a worst case was slow enough to cause a timeout error,
causing the test to fail/flake.

Note that we must at least have a query other than "*", though. If the
query is simply "*", then we won't retry misspellings, anyway, so the
test isn't meaningful in that case.